rename programInterface to programPanel
